Correct ax registers
The order was wrong, as per https://github.com/dolphin-emu/dolphin/commit/d5c3411a8054e45557f406f080732270f5427add.  Furthermore, the size should be 4, not 6; I had it as 6 under the impresion that there was an instruction that used either an ac0/1 or an ax0/1 register that required the sizes to match, but that wasn't the case.