Fix immediate values
The const special address space is needed to properly determine the size.  Not using it will result in a size-0 value being used, which is always zero.  There is a warning for this, but it's not super clear what needs to be done about it.

One other note: Loops actually work properly now, but you need to delete and re-create a project.  Otherwise, stale context data will still exist, and ghidra will do the wrong thing, even if you clear and re-disassemble.