Fix SUBARN
The previous encoding conflicted with ADDARN, meaning that ADDARN $arD, $ix3 would instead be interpretted as SUBARN $arD, $ixD (while the actual SUBARN instruction was treated as invalid).

There was a second bug where ix0_2 wasn't associated with any registers, meaning it was instead treated as a number, so SUBARN $ar2, $ix2 was instead parsed (and emulated as) SUBARN $ar2, #2. This second bug also affected ILRRN.