Rework loop handling
Now, the loop skip and/or jumpback address are shown as operands, which fixes Ghidra treating the instruction 2 after a LOOP instruction as the start of a new function, and also improves xref behavior.

However, it now also treats the entire set of parameters to the instruction (both the main and extended one) as a single operand, which makes other behavior a bit weirder. I think this is still a net improvement, though.

Note that you will need to run the ReloadSleighLangauge.java script after updating, and then re-perform analysis (by pressing A), for xrefs to properly be created/restored.